# Explicitly mark get_precise_ns on OSX as inline.