--- a/benches/precise_time_ns.rs
+++ b/benches/precise_time_ns.rs
@@ -7,5 +7,8 @@ use test::Bencher;
 
 #[bench]
 fn bench_precise_time_ns(b: &mut Bencher) {
-    b.iter(|| time::precise_time_ns())
+    b.iter(|| {
+        time::precise_time_ns();
+        time::precise_time_ns();
+    });
 }