Improve performance on OSX ~1.5-2x:

before 61 ns/iter (+/- 9)
after 43 ns/iter (+/- 11)

Not as visible benchmarking a single call but much more noticeable with many
calls.  For example for 10 calls it's:

before 328 ns/iter (+/- 54)
after 187 ns/iter (+/- 53)

Thus bump the benchamrk to do 2 calls to precise_time_ns.

Fixes #159